Created redirect img component & organized assets
