+++ bloom-and-vine/src/Components/redirect-display/components/redirect-image/RedirectImage.css
@@ -1,15 +1,3 @@
-.image-container {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    flex-wrap: nowrap;
-    gap: 103px;
-
-    margin: 116px 0px 153px 0px;
-
-}
-
 .image-background {
     position: relative;
     min-width: 257px;
@@ -44,15 +32,15 @@
     position: absolute;
     object-fit: cover;
     width: 256px;
-    height: 276px;
+    max-height: 276px;
     top: -72px;
     left: -42px;
 
     filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
     border-radius: 10px;
+    transition: 0.5s;
 }
 
 .image-redirect:hover{
     transform: scale(1.15);
-    transition: 0.5s;
 }
