+++ bloom-and-vine/src/Components/review-slider/ReviewSlider.css
@@ -13,6 +13,8 @@ button:hover {
 .review-slider {
     display: flex;
     flex-direction: row;
+    flex-wrap: nowrap;
     align-items: center;
     justify-content: center;
+    
 }
