+++ bloom-and-vine/src/Components/review-slider/components/user-review/UserReview.css
@@ -2,6 +2,7 @@
     position: relative;
     display: flex;
     flex-direction: column;
+    flex-wrap: wrap;
     align-items: center;
     padding: 100px;
 }
@@ -40,8 +41,5 @@
     align-items: center;
     justify-content: center;
     flex-wrap: nowrap;
-}
-
-.review-dot > *:last-child {
-    padding-right: 16px;
+    gap: 16px;
 }
