+++ bloom-and-vine/src/Components/review-slider/components/user-review/UserReview.js
@@ -1,23 +1,20 @@
-import PurpleCircle from './logos/purple_circle.png'
-import WhiteCircle from './logos/white_circle.png'
+import PurpleCircle from '../../assets/purple_circle.png'
+import WhiteCircle from '../../assets/white_circle.png'
 import './UserReview.css'
 
-export default function UserReview(props) {
-    const { review, curReview, totalReviews } = props;
+function fetchDots(curIndex, totalDots) {
+    const dots = []
 
-    let dot_display = []
-
-    for (let i = 0; i < totalReviews; i++) {
-        if (i !== curReview) {
-            dot_display.push(
+    for (let i = 0; i < totalDots; i++) {
+        if (i !== curIndex) {
+            dots.push(
                 <li className='review-dot'>
                     <img src={WhiteCircle}/>
                 </li>
             )
         }
-
         else {
-            dot_display.push(
+            dots.push(
                 <li className='review-dot'>
                     <img src={PurpleCircle}/>
                 </li>
@@ -25,6 +22,13 @@ export default function UserReview(props) {
         }
     }
 
+    return dots
+}
+
+export default function UserReview(props) {
+    const { review, curReview, totalReviews } = props;
+
+    const dot_display = fetchDots(curReview, totalReviews)
 
     return (
         <div className='review-container'>
